:package: moved interfaces to separate file
* also added some folders
